Added prop to accept other mime types
Currently the only mime type that can be set is 'image/*', that is set with the image prop. Now the field allows any mime type and if both, accept and image props are set, accept prop is going to stay. This way you can choose image and also accept only jpg images, for example.